+++ components/openapi/src/it/scala/pl/touk/nussknacker/openapi/functional/OpenAPIServiceSpec.scala
@@ -32,7 +32,7 @@ class OpenAPIServiceSpec
 
   implicit val componentUseCase: ComponentUseCase = ComponentUseCase.EngineRuntime
   implicit val metaData: MetaData                 = MetaData("testProc", StreamMetaData())
-  implicit val contextId: Context                 = Context("testContextId", Map.empty)
+  implicit val context: Context                   = Context("testContextId", Map.empty)
 
   type FixtureParam = EagerServiceWithStaticParametersAndReturnType