+++ designer/restmodel/src/test/scala/pl/touk/nussknacker/restmodel/process/ProcessPropertiesTest.scala
@@ -4,7 +4,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks.forAll
 import pl.touk.nussknacker.engine.api.{ProcessAdditionalFields, TypeSpecificData}
-import pl.touk.nussknacker.engine.api.MetaDataTestData.{fullMetaDataCases, nonTypeSpecificProperties}
+import pl.touk.nussknacker.engine.api.TypeSpecificDataTestData.{fullMetaDataCases, nonTypeSpecificProperties}
 import pl.touk.nussknacker.engine.api.graph.ProcessProperties
 import pl.touk.nussknacker.engine.api.process.ProcessName