+++ engine/flink/components/base/src/main/scala/pl/touk/nussknacker/engine/flink/util/transformer/FlinkBaseComponentProvider.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.engine.flink.util.transformer
 import com.typesafe.config.Config
 import pl.touk.nussknacker.engine.api.component.{ComponentDefinition, ComponentProvider, NussknackerVersion}
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, SinkFactory}
+import pl.touk.nussknacker.engine.common.components.DecisionTable
 import pl.touk.nussknacker.engine.flink.util.sink.EmptySink
 import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.AggregateWindowsConfig
 import pl.touk.nussknacker.engine.flink.util.transformer.aggregate.sampleTransformers.{
@@ -43,6 +44,10 @@ object FlinkBaseComponentProvider {
       ComponentDefinition("union", UnionTransformer).withRelativeDocs("BasicNodes#union"),
       ComponentDefinition("dead-end", SinkFactory.noParam(EmptySink)).withRelativeDocs("DataSourcesAndSinks#deadend"),
       ComponentDefinition("periodic", PeriodicSourceFactory).withRelativeDocs("DataSourcesAndSinks#periodic"),
+      ComponentDefinition(
+        name = "decision-table",
+        component = DecisionTable
+      ).withRelativeDocs("BasicNodes#decisiontable")
     )
 
     val statefulComponents = List(
@@ -64,11 +69,7 @@ object FlinkBaseComponentProvider {
       ComponentDefinition("full-outer-join", FullOuterJoinTransformer).withRelativeDocs(
         "AggregatesInTimeWindows#full-outer-join"
       ),
-      ComponentDefinition("delay", DelayTransformer).withRelativeDocs("DataSourcesAndSinks#delay"),
-//      ComponentDefinition(
-//        name = "decision-table",
-//        component = DecisionTable
-//      ).withRelativeDocs("BasicNodes#decisiontable")
+      ComponentDefinition("delay", DelayTransformer).withRelativeDocs("DataSourcesAndSinks#delay")
     )
 
     statefulComponents ++ statelessComponents